src/app/feature/conductor/components/crear-conductor/crear-conductor.component.ts
import { Component, OnInit } from '@angular/core';
import { FormGroup, FormControl, Validators } from '@angular/forms';
import { ConductorService } from '../../shared/service/conductor.service';

const LONGITUD_MINIMA_PERMITIDA_TEXTO = 3;
const LONGITUD_MAXIMA_PERMITIDA_TEXTO = 20;

@Component({
  selector: 'app-crear-conductor',
  templateUrl: './crear-conductor.component.html',
  styleUrls: ['./crear-conductor.component.css']
})
export class CrearConductorComponent implements OnInit {
  conductorForm: FormGroup;
  constructor(protected conductorService: ConductorService) { }

  ngOnInit() {
    this.construirFormularioConductor();
  }

  crear() {
    this.conductorService.guardar(this.conductorForm.value);
  }

  private construirFormularioConductor() {
    this.conductorForm = new FormGroup({
      id: new FormControl('', [Validators.required]),
      descripcion: new FormControl('', [Validators.required, Validators.minLength(LONGITUD_MINIMA_PERMITIDA_TEXTO),
                                                             Validators.maxLength(LONGITUD_MAXIMA_PERMITIDA_TEXTO)])
    });
  }

}
